requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
tensorflow-cpu==2.20.0
numpy==1.26.2
Pillow==10.1.0
